Removed the fallback in the editor launch
The editor is a vital part of noScribe, so I don't want the launch to fail silently, using the browser as a fallback. It now throws an error if noScribeEdit is not found.
Second small change: `program: str | None = None` seems to be python 3.10 syntax. Changed it to `program: str = None`